refactor(tests): rewrite tests to ava
Closes #65.
