+++ tests/__helpers__/testRule.js
@@ -1,0 +1,31 @@
+const { readFile } = require( 'node:fs/promises' );
+const test = require( 'ava' );
+const { ESLint } = require( 'eslint' );
+const defaultConfig = require( '../../index.js' );
+
+const testRule = test.macro( async ( t, {
+	fixturePath,
+	config = defaultConfig,
+	fakePath = fixturePath,
+	errorCount = 0
+} = {} ) => {
+	const  [ results ] = await lintFile( fixturePath, config, fakePath );
+
+	t.is( results.errorCount, errorCount );
+} );
+
+async function lintFile( path, config, fakePath = path ) {
+	const code = await readFile( path, 'utf8' );
+	const eslint = new ESLint( {
+		ignore: false,
+		useEslintrc: false,
+		baseConfig: config
+	} );
+	const report = await eslint.lintText( code, {
+		filePath: fakePath
+	} );
+
+	return report;
+}
+
+module.exports = testRule;
